pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.15"

dependencies = [
    "aiohttp>=3.13.5",
    "av>=12.0.0",
    "httpx>=0.28.1",
    "livekit-agents[silero,turn-detector]~=1.5",
    "livekit-plugins-ai-coustics~=0.2",
    "livekit-plugins-deepgram>=1.5.6",
    "livekit-plugins-elevenlabs~=1.5",
    "livekit-plugins-openai>=1.5.6",
    "python-dotenv",
    "rapidfuzz>=3.14.5",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-xdist>=3.8.0",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
